A whole bunch of messy attempts to getting @Preview working in Desktop. Must remember that Preview doesn't work for functions that take parameters.
