+++ androidApp/build.gradle.kts
@@ -10,6 +10,7 @@ kotlin {
         val androidMain by getting {
             dependencies {
                 implementation(project(":shared"))
+                implementation(compose.preview)
             }
         }
     }
