+++ desktopApp/src/jvmMain/kotlin/main.kt
@@ -1,4 +1,8 @@
+import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
@@ -22,4 +26,10 @@ fun main() = application {
         )
         DesktopApp(isDark = dark)
     }
+}
+
+@Preview
+@Composable
+fun Wibble() {
+    Text(text = "Wibble", color = Color.Blue)
 }
