+++ shared/src/commonMain/kotlin/org.liamjd.cantilever.corbel/ui/models/TabBar.kt
@@ -1,0 +1,8 @@
+package org.liamjd.cantilever.corbel.ui.models
+
+enum class Tabs {
+    POSTS,
+    PAGES,
+    TEMPLATES
+}
+
